+++ src/app/Fake.IO.FileSystem/Globbing.fs
@@ -1,5 +1,6 @@
 /// This module contains a file pattern globbing implementation.
 /// This module is part of the `Fake.IO.FileSystem` package
+[<RequireQualifiedAccess>]
 module Fake.IO.Globbing.Glob
 
 open System